move nbformat up by itself
