+++ src/py/jupyter_graphql/schema/contents/base.py
@@ -5,11 +5,3 @@ def CM(info):
 
 def RESOLVE_CONTENT(it, info):
     return it["content"] or CM(info).get(it["path"])["content"]
-
-
-def GET(attr):
-    return lambda it, info: it.get(attr)
-
-
-def IDENTITY(it, info):
-    return it
